Python 3.6 support
Python 3.6+ now fails on invalid back references.  These would only
appear in comments in verbose mode as we *should* handle all other
cases and convert them appropriately (unless the user is using them
improperly).  It is okay if Re asserts on a user using them improperly,
but it shouldn’t fail on back references in comments.  The solution is
just to escape *our* back references in comments in verbose mode.

fixes #2