Check lint on PY3
Move to checking lint on PY3.  In 2020 PY2 will be end of life, so move to PY3 now.  Also add linter to catch improper use of mutables and also override of builtins.  Ignore current builtin errors as we mirror naming of argmunts of regex modules we are providing a wrapper for.