Unicode Names (#19)
* Initial unicode name work (WIP)

Add support for \N in search and replace.  Add logic to catch when references like \N, \p, \P, and \g are incomplete and raise an exception. At a little testing (still more to do).

* Update doc theme and config to fix failure

* Ignores Py37 and remove Py33

* Run Python 3.7 *and* allow failure

* Remove Python 3.3 from appveyor test

* Coverage 100% with small fixes

Properly access the patterns named Unicode value and add test remaining cases.

* Python2 chokes on incomplete Unicode test

This occurs during compile which makes it impossible to stub out conditionally.  Probably need to have a PY3 specific test file for this and only run it on PY3, but for now, let's just skip the test.