Better string format parsing
Reduce format passes.  Handle index integer conversion and regex sub-capture injection on intial validation pass.  Save field_name in get_format and pass it to handle_format_group so we don't have to reparse again. Also, hopefully fix PY2 handling as the fieldname splitter is a method of the actual string object, and not a separate function like it is in PY3.