Deprecate boost style case references
For consistency between re and regex libraries, \l, \L, \c, and \C will
be removed from the bre library. This is the first step: deprecation.
The POSIX form, which these references where shortcuts for, should be
used instead. The removal only pertains to search patterns. These are
still available in both re and regex replace patterns for augmenting
case of characters.